WIP send response: output raw packet is good but stuffed packet is null. CRC with one byte null fixed: crc_t typedef was set to uint8_t instead of uint16_t

+++ src/constants.h
@@ -39,6 +39,6 @@
  * The width of the CRC calculation and result.
  * Modify the typedef for a 16 or 32-bit CRC standard.
  */
-typedef uint8_t crc_t;
+typedef uint16_t crc_t;
 
 #endif /* CONSTANTS_H_ */
